0.0.1

+++ res/resize.sh
@@ -5,15 +5,19 @@ cp icon512x512.png ../src/img/icon.png
 cp icon512x512.png ../src/img/128.png
 cp icon512x512.png ../src/img/48.png
 cp icon512x512.png ../src/img/16.png
-cp icon512x512-disabled.png ../src/img/icon-disabled.png
-cp icon512x512-disabled.png ../src/img/128-disabled.png
-cp icon512x512-disabled.png ../src/img/48-disabled.png
-cp icon512x512-disabled.png ../src/img/16-disabled.png
+if [ -e "icon512x512-disabled.png" ]; then
+  cp icon512x512-disabled.png ../src/img/icon-disabled.png
+  cp icon512x512-disabled.png ../src/img/128-disabled.png
+  cp icon512x512-disabled.png ../src/img/48-disabled.png
+  cp icon512x512-disabled.png ../src/img/16-disabled.png
+fi
 pushd ../src/img
 sips -Z 128 128.png
 sips -Z 48 48.png
 sips -Z 16 16.png
-sips -Z 128 128-disabled.png
-sips -Z 48 48-disabled.png
-sips -Z 16 16-disabled.png
+if [ -e "icon512x512-disabled.png" ]; then
+  sips -Z 128 128-disabled.png
+  sips -Z 48 48-disabled.png
+  sips -Z 16 16-disabled.png
+fi
 popd
